requirements.txt
openai==0.28
ffmpeg-python
openai-whisper
streamlit
pytube